pyproject.toml
[project]
name = "trace-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.86.0",
    "awslabs-mcp-lambda-handler>=0.1.14",
    "mangum>=0.21.0",
    "mcp[cli]>=1.26.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]
